pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_project"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "DavidLindahl", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "dvc[gs]>=3.66.1",
    "python-multipart>=0.0.9",
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "kaggle>=1.8.3",
    "matplotlib>=3.10.8",
    "numpy>=1.25.2",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "ruff>=0.1.3",
    "tqdm>=4.66.5",
    "timm>=1.0.24",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "typer==0.21.1",
    "uvicorn==0.34.0",
    "pip>=25.3",
    "google-api-python-client>=2.188.0",
    "dvc-gs>=3.0.2",
    "wandb>=0.24.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "coverage==7.13.1",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
    "pillow>=12.1.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cuda", marker = "sys_platform != 'darwin'" },
]
torchvision = [
    { index = "pytorch-cuda", marker = "sys_platform != 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[project.scripts]
train = "mlops_project.train:main"
data = "mlops_project.data:main"
preprocess = "mlops_project.preprocess:main"
inference = "mlops_project.inference:app"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
omit = ["tests/*"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_configs = true
